LLEXT: Cleanup and moving around files.
