+++ app/src/applications/calculator/calculator_app.c
@@ -7,7 +7,7 @@
 
 #include "calculator_ui.h"
 #include "managers/zsw_app_manager.h"
-#include "managers/zsw_llext_iflash.h"
+#include "llext/zsw_llext_iflash.h"
 #include "ui/utils/zsw_ui_utils.h"
 #include "smf_calculator_thread.h"
