+++ app/src/applications/qr_code/qr_code_app.c
@@ -21,7 +21,7 @@
 
 #include "qr_code_ui.h"
 #include "managers/zsw_app_manager.h"
-#include "managers/zsw_llext_iflash.h"
+#include "llext/zsw_llext_iflash.h"
 #include "ui/utils/zsw_ui_utils.h"
 
 #ifdef CONFIG_ZSW_LLEXT_APPS
