+++ app/src/applications/trivia/trivia_app.c
@@ -28,7 +28,7 @@
 
 #include "trivia_ui.h"
 #include "managers/zsw_app_manager.h"
-#include "managers/zsw_llext_iflash.h"
+#include "llext/zsw_llext_iflash.h"
 #include "ui/utils/zsw_ui_utils.h"
 #include <ble/ble_http.h>
 #include "cJSON.h"
