+++ app/src/llext/CMakeLists.txt
@@ -11,5 +11,8 @@ if(CONFIG_ZSW_LLEXT_APPS)
     target_sources(app PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/zsw_llext_exports.c
         ${CMAKE_CURRENT_SOURCE_DIR}/zsw_llext_log.c
+        ${CMAKE_CURRENT_SOURCE_DIR}/zsw_llext_app_manager.c
+        ${CMAKE_CURRENT_SOURCE_DIR}/zsw_llext_xip.c
+        ${CMAKE_CURRENT_SOURCE_DIR}/zsw_llext_iflash.c
     )
 endif()
